Updated Maze.java, Animate.java, ChapTile.java
Removed the enum from Animate, as it is more productive to store in Maze

Animate now is not being used, though can be still ported 
over to be used from changes 
made in the other files (Saving Space at the moment.

I updated the moveTile class to accepted a direction 
parameter from an enum I ported from animate to Maze.

made all moveDIRECTION(); have a direction from the enum

I made new methods in ChapTile to change the filename 
according to the direction in moveTile ( in Maze.java )

Animation should be seen with chap going left to right